Added factory for default values
The factories NewConfigLxcFromJson/NewConfigLxcFromApi create ConfigLxc
structures that have some fields unset, for example having "0" cores.

The code from NewConfigQemuFromApi solved this by initializing a default
value for each member as a local variable, then checking if it was
defined in the map gotten from PVEAPI, then setting a new value, then
assigning that variable to a member of the new construct. In the case of
NewConfigQemuFromJson, this was left to Decoder.Decode() to decide.

This is actually job for a base factory that sets every field to
a default value, the defaults are pick from the PVEAPI docs. This also
trimmed the double assignment in NewConfigLxcFromApi.

Also fixed some typos in the struct definition