Prepare code in Client for LXC creation
CreateLxcVm , needed to pass "lxc" in the url. This is the only change
wrt to CreateQemuVm, so it will be parameterised later

CreateVMDisks, adapted to create the mountpoints for containers if they
appear in the configuration. Mind that such volumes will make CT
creation to fail because they won't contain a filesystem. Don't use
mountpoints until filesystem creation on the fly is figured out.