Working ConfigLxc.UpdateConfig
This method exposed a problem in NewConfig(Qemu/Lxc)FromJson, stemming in
that they deserialize to a struct with annotations instead of to
something more generic like a map[string]interface{} or interface{}

Details in the comments in the method.

Also added a check in CreateLxcMpParams so when rootfs isn't set in
the user supplied JSON (useful for UpdateConfig), it won't be created
again.