Added CT Clone with new JSON submap
Instead of using two different client methods with two different Go
structures each containing a fixed number of members for the parameters
that differ between the two PVEAPI cloning operations, just add
a "clone" key in the json config and a CloneParams map[string]interface{}
that can be passed directly to a single method (integrated in the
previous commit).

Same approach will be used for ConfigQemu.CloneVm, as this also brings
the added benefit of being able to use all the parameters available in
PVEAPI without needing to hardcode anything.