Move clone code to Client
While editing ConfigQemu and ConfigLxc CloneVm methods, it got to
a point where they did three things: set the vmtype (for the new vmref) to
"qemu" or "lxc", calling Client.CloneVm, and then calling the relevant
UpdateConfig method.

Setting the vmtype can be done from a central point. But this is not
even needed: PVEAPI doesn't need to infer that the clone will be of the
same type of the source, it is implicit.

Calling UpdateConfig breaks single responsibility. To update the clone
config, apply UpdateConfig to it after cloning.

After trimming everything there are no more need for specific methods
for this operation, or to mix a JSON substructure into the VM/CT
configurations and data structures for cloning information.

The Client.CloneVm method expects a JSON string with each member setting
options as per PVEAPI doc for VM or CT cloning (there are some options
that are different). This can be handled in agnostic fashion by passing
a map[string]interface{} instead of a struct to Decoder.Decode