Move VM related code out of Client
Move all the Vm code to its own file.

Same is done with the function GetNodeList for a new node unit, for now
just leave it in its own file, later Node related code can be there.

For now the client is passed as a parameter for all the functions and
methods that need it, but this will change in the next commits to have
a global GetClient function that gets the client in use.