Remove client parameter passing
Replaced explicit passing with:

var setClient *Client

Stores the selected client to use in calls

(*Client) Set()

Sets the client to use in setClient

GetClient() *Client

Returns the selected client.

User code should set a client as soon as it is created and before
calling code that might need it. The library code then calls GetClient
to perform Client requests.